refactor(ai-worker): replica estrutura wireframe na etapa presetdesign
