+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/presetdesign/monitor/PresetDesignPendingJobsService.java
@@ -1,6 +1,6 @@
 package com.marketinghub.worker.geralanding.presetdesign.monitor;
 
-import com.marketinghub.worker.geralanding.presetdesign.GeraLandingPresetDesignBackendClient;
+import com.marketinghub.worker.geralanding.presetdesign.backend.GeraLandingPresetDesignBackendClient;
 import com.marketinghub.worker.geralanding.presetdesign.dto.GeraLandingStageExecutionDetailDto;
 import java.util.List;
 import java.util.Locale;
