+++ ai-worker/src/main/java/com/marketinghub/worker/geralanding/presetdesign/response/RecebeResponse.java
@@ -1,4 +1,4 @@
-package com.marketinghub.worker.geralanding.presetdesign;
+package com.marketinghub.worker.geralanding.presetdesign.response;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,16 +13,16 @@ public class RecebeResponse {
 
     private static final Logger log = LoggerFactory.getLogger(RecebeResponse.class);
 
-    private final GeraLandingPresetDesignBackendClient backendClient;
+    private final com.marketinghub.worker.geralanding.presetdesign.backend.GeraLandingPresetDesignBackendClient backendClient;
 
-    public RecebeResponse(GeraLandingPresetDesignBackendClient backendClient) {
+    public RecebeResponse(com.marketinghub.worker.geralanding.presetdesign.backend.GeraLandingPresetDesignBackendClient backendClient) {
         this.backendClient = backendClient;
     }
 
     /**
      * Envia para o backend os dados de despacho e de resultado da etapa, incluindo a resposta crua da OpenAI.
      */
-    public void processar(Long experimentId, String stageCode, String idJob, GeraLandingJobCompletionPresetDesignPayload payload) {
+    public void processar(Long experimentId, String stageCode, String idJob, RecordPresetDesignResponse payload) {
         if (payload != null && payload.openAiJobId() != null && !payload.openAiJobId().isBlank()) {
             backendClient.receiveDispatch(idJob, experimentId, stageCode, payload.openAiJobId());
         }
